+++ locale/eu-eu/vocabulary/change.wallpaper.intent
@@ -1,0 +1,3 @@
+aldatu horma papera
+aldatu pantaila nagusia
+aldatu pantaila
