+++ locale/eu-eu/vocabulary/wallpaper-error.dialog
@@ -1,0 +1,1 @@
+Errorea gertatu da horma-papera aldatzean. Ziurtatu URLa jaitsi daitekeen irudi batena dela.
